[FIX] mail: correctly compute `direct_partner_id` for dm_chat
Revision on https://github.com/odoo/odoo/commit/9ca3fcfc707816e702fc67397cd730c8d22c5ddf

The commit above improves the channel_info, so that there
are less DB queries.

However, the computation was wrong for partners on dm chat,
which resulted in omitting the name and email of the partner.

closes odoo/odoo#31033