[IMP] gamification,forum: remove deprecated alert when sending badge
Since 129fd7587b88b3e87cc022e13ac29ccf9c18a787, when a user get a badge,
an email is sent. The subtype is not used anymore, so the alert
mecanism in forum to display a message to the user to show him
his new badges can not work.
This commit removes this dead code.
Task-1922159