[FIX] mail: notify channel_fetched/channel_seen only when changes
Revision on https://github.com/odoo/odoo/commit/07e6dbdb9410a917cb14d870e78340fa2f289364

With the commit above, when the JS notified on `channel_fetched` and
`channel_seen`, it was always writing in DB and sending notifications
on the longpoll.

This is useless if the information on last fetched/seen message
of the conversation remains unchanged, which is what this commit
solves.

Also prevent `channel_fetched` from seding notifications on
the longpoll if this is called on a non-chat channel.