[FIX] web_editor: fix switch design (website publish btn)
Using chrome, the design was not pixel perfect (letting a white bar
appear around the switch). Replacing the linear gradient color by a
simple color fixes that particular problem. The design is still not
pixel perfect but those are meant to be replaced in master anyway.

Part of https://github.com/odoo/odoo/pull/31117
task-1941156